cstack: add video export to CorrelatedStack
Because the APIs are now comparable, video export can simply be spun out of `Scan` and used in both `Scan` and `CorrelatedStack`.
